- add sceneObject scene handling - add scene ID's to SceneCreatorModule - add object caching to UpdateReceiverModule - temporary disable keying shader - add factory to SceneObject to realize init with parameters

+++ VPET_Unity2/Assets/VPET/Core/Core.cs
@@ -376,7 +376,7 @@ namespace vpet
             {
                 Dictionary<short, ParameterObject> sceneObjects;
                 if (m_parameterObjectList.TryGetValue(sceneID, out sceneObjects))
-                    return sceneObjects[(short)(poID)];
+                    return sceneObjects[poID];
                 else
                     return null;
             }
